requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
requests
pyarrow
db-dtypes
